+++ docs/CHANGELOG.md
@@ -15,6 +15,18 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.85.3] - 01-07-24
+### Fixed
+- Another issue with docker release.
+
+## [0.85.2] - 01-07-24
+### Fixed
+- Issues with docker release.
+
+## [0.85.1] - 01-07-24
+### Fixed
+- Bug when using the `get_cognite_client` function with interactive login. This is now fixed.
+
 ## [0.85.0] - 25-06-24
 ### Changed
 - [BREAKING] Interface for `Loaders`. Instead of `.export_to_cdf` now always return `UploadResultList` and